Refactor reconciler work builder
