+++ catalogue_graph/tests/adapters/transformers/marc/test_transformer.py
@@ -5,11 +5,11 @@ from datetime import datetime
 from typing import Any, cast
 
 import pytest
+from elasticsearch import Elasticsearch
+
 from adapters.utils.adapter_store import AdapterStore
 from core.source import BaseSource
-from elasticsearch import Elasticsearch
 from models.pipeline.source.work import VisibleSourceWork
-
 from tests.adapters.transformers.marc.marcxml_test_transformer import (
     MarcXmlTransformerForTests,
 )
